migrates session
