+++ itu-minitwit-node/src/db/session.js
@@ -1,0 +1,56 @@
+const session = require('express-session');
+const MySQLStore = require('express-mysql-session')(session);
+require('dotenv').config();
+
+const options = {
+  host: process.env.MYSQL_HOST || 'localhost',
+  port: process.env.MYSQL_PORT || 3306,
+  user: process.env.MYSQL_USERNAME || 'root',
+  password: process.env.MYSQL_PASSWORD || 'root',
+  database: process.env.MYSQL_DATABASE || 'defaultdb',
+  // Whether or not to automatically check for and clear expired sessions:
+  clearExpired: true,
+  // How frequently expired sessions will be cleared; milliseconds:
+  checkExpirationInterval: 900000,
+  // The maximum age of a valid session; milliseconds:
+  expiration: 86400000,
+  // Whether or not to create the sessions database table, if one does not already exist:
+  createDatabaseTable: true,
+  // Whether or not to end the database connection when the store is closed.
+  // The default value of this option depends on whether or not a connection was passed to the constructor.
+  // If a connection object is passed to the constructor, the default value for this option is false.
+  endConnectionOnClose: true,
+  // Whether or not to disable touch:
+  disableTouch: false,
+  charset: 'utf8mb4_bin',
+  schema: {
+    tableName: 'sessions',
+    columnNames: {
+      session_id: 'session_id',
+      expires: 'expires',
+      data: 'data'
+    }
+  }
+};
+
+const sessionStore = new MySQLStore(options);
+
+// Optionally use onReady() to get a promise that resolves when store is ready.
+sessionStore.onReady().then(() => {
+	// MySQL session store ready for use.
+	console.log('MySQLStore ready');
+}).catch(error => {
+	// Something went wrong.
+	console.error(error);
+});
+
+sessionStore.close().then(() => {
+	// Successfuly closed the MySQL session store.
+	console.log('MySQLStore closed');
+}).catch(error => {
+	// Something went wrong.
+	console.error(error);
+});
+
+
+module.exports = sessionStore;
